chore: merge full DID utils into full DID
