+++ jest.config.js
@@ -34,7 +34,6 @@ module.exports = {
     '!did/src/Did.chain.ts',
     '!did/src/Did.utils.ts',
     '!did/src/DidDetails/LightDidDetails.utils.ts',
-    '!did/src/DidDetails/FullDidDetails.utils.ts',
     '!utils/src/jsonabc.ts',
   ],
   resolver: "ts-jest-resolver",
