+++ packages/did/src/DidDetails/FullDidDetails.spec.ts
@@ -320,3 +320,78 @@ describe('When creating an instance from the chain', () => {
     })
   })
 })
+
+const mockApi = ApiMocks.createAugmentedApi()
+
+describe('When creating an instance from the chain', () => {
+  it('Should return correct KeyRelationship for single valid call', () => {
+    const keyRelationship = Did.getKeyRelationshipForExtrinsic(
+      mockApi.tx.attestation.add(new Uint8Array(32), new Uint8Array(32), null)
+    )
+    expect(keyRelationship).toBe('assertionMethod')
+  })
+  it('Should return correct KeyRelationship for batched call', () => {
+    const keyRelationship = Did.getKeyRelationshipForExtrinsic(
+      mockApi.tx.utility.batch([
+        mockApi.tx.attestation.add(
+          new Uint8Array(32),
+          new Uint8Array(32),
+          null
+        ),
+        mockApi.tx.attestation.add(
+          new Uint8Array(32),
+          new Uint8Array(32),
+          null
+        ),
+      ])
+    )
+    expect(keyRelationship).toBe('assertionMethod')
+  })
+  it('Should return correct KeyRelationship for batchAll call', () => {
+    const keyRelationship = Did.getKeyRelationshipForExtrinsic(
+      mockApi.tx.utility.batchAll([
+        mockApi.tx.attestation.add(
+          new Uint8Array(32),
+          new Uint8Array(32),
+          null
+        ),
+        mockApi.tx.attestation.add(
+          new Uint8Array(32),
+          new Uint8Array(32),
+          null
+        ),
+      ])
+    )
+    expect(keyRelationship).toBe('assertionMethod')
+  })
+  it('Should return correct KeyRelationship for forceBatch call', () => {
+    const keyRelationship = Did.getKeyRelationshipForExtrinsic(
+      mockApi.tx.utility.forceBatch([
+        mockApi.tx.attestation.add(
+          new Uint8Array(32),
+          new Uint8Array(32),
+          null
+        ),
+        mockApi.tx.attestation.add(
+          new Uint8Array(32),
+          new Uint8Array(32),
+          null
+        ),
+      ])
+    )
+    expect(keyRelationship).toBe('assertionMethod')
+  })
+  it('Should return undefined for batch with mixed KeyRelationship calls', () => {
+    const keyRelationship = Did.getKeyRelationshipForExtrinsic(
+      mockApi.tx.utility.forceBatch([
+        mockApi.tx.attestation.add(
+          new Uint8Array(32),
+          new Uint8Array(32),
+          null
+        ),
+        mockApi.tx.web3Names.claim('awesomename'),
+      ])
+    )
+    expect(keyRelationship).toBeUndefined()
+  })
+})
